# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Multitenant SaaS Appointment Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business/profile:
    get:
      operationId: getBusinessProfile
      tags: [business]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBusinessProfile
      tags: [business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/branding:
    patch:
      operationId: updateBusinessBranding
      tags: [business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrandingBody"
      responses:
        "200":
          description: Branding updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/integrations:
    patch:
      operationId: updateBusinessIntegrations
      tags: [business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIntegrationsBody"
      responses:
        "200":
          description: Integrations updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/services:
    get:
      operationId: listBusinessServices
      tags: [business]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBusinessService
      tags: [business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/services/{id}:
    patch:
      operationId: updateBusinessService
      tags: [business]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBusinessService
      tags: [business]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/working-hours:
    get:
      operationId: getWorkingHours
      tags: [business]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Working hours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingHour"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: setWorkingHours
      tags: [business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetWorkingHoursBody"
      responses:
        "200":
          description: Working hours updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingHour"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/break-times:
    get:
      operationId: getBreakTimes
      tags: [business]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Break times
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BreakTime"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: setBreakTimes
      tags: [business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBreakTimesBody"
      responses:
        "200":
          description: Break times updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BreakTime"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/appointments:
    get:
      operationId: listBusinessAppointments
      tags: [business]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Appointments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/appointments/{id}:
    delete:
      operationId: cancelBusinessAppointment
      tags: [business]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/stats:
    get:
      operationId: getBusinessStats
      tags: [business]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/customers:
    get:
      operationId: listBusinessCustomers
      tags: [business]
      summary: CRM - list unique clients with history and revenue
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/waitlist:
    get:
      operationId: listBusinessWaitlist
      tags: [business]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Waitlist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaitlistEntry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /business/waitlist/{id}:
    delete:
      operationId: removeFromWaitlist
      tags: [business]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Public Client Booking ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    BusinessProfile:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        ownerName:
          type: string
        email:
          type: string
        bufferMinutes:
          type: integer
        notificationEnabled:
          type: boolean
        notificationMessage:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        fontFamily:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        themeMode:
          type: ["string", "null"]
          enum: ["light", "dark", null]
        whatsappApiKey:
          type: ["string", "null"]
        whatsappPhoneId:
          type: ["string", "null"]
        googleCalendarEnabled:
          type: boolean
        stripeEnabled:
          type: boolean
        stripePublicKey:
          type: ["string", "null"]
        subscriptionPlan:
          type: string
        maxServicesAllowed:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - slug
        - name
        - ownerName
        - email
        - bufferMinutes
        - notificationEnabled
        - notificationMessage
        - primaryColor
        - fontFamily
        - logoUrl
        - bannerUrl
        - themeMode
        - whatsappApiKey
        - whatsappPhoneId
        - googleCalendarEnabled
        - stripeEnabled
        - stripePublicKey
        - subscriptionPlan
        - maxServicesAllowed
        - isActive
        - createdAt
    UpdateBusinessProfileBody:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: string
        bufferMinutes:
          type: integer
        notificationEnabled:
          type: boolean
        notificationMessage:
          type: ["string", "null"]
    UpdateBrandingBody:
      type: object
      properties:
        primaryColor:
          type: ["string", "null"]
        fontFamily:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        themeMode:
          type: ["string", "null"]
    UpdateIntegrationsBody:
      type: object
      properties:
        whatsappApiKey:
          type: ["string", "null"]
        whatsappPhoneId:
          type: ["string", "null"]
        googleCalendarEnabled:
          type: boolean
        stripeEnabled:
          type: boolean
        stripePublicKey:
          type: ["string", "null"]
...
    Service:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        price:
          type: integer
        durationMinutes:
          type: integer
        bufferMinutes:
          type: integer
          description: Buffer time after this service
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, businessId, name, price, durationMinutes, bufferMinutes, imageUrl, isActive, createdAt]
    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
        durationMinutes:
          type: integer
        bufferMinutes:
          type: integer
          description: Buffer time after this service in minutes
        imageUrl:
          type: ["string", "null"]
      required: [name, price, durationMinutes]
    UpdateServiceBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
        durationMinutes:
          type: integer
        bufferMinutes:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    WorkingHour:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        isEnabled:
          type: boolean
      required: [id, businessId, dayOfWeek, startTime, endTime, isEnabled]
    SetWorkingHoursBody:
      type: object
      properties:
        hours:
          type: array
          items:
            $ref: "#/components/schemas/WorkingHourInput"
      required: [hours]
    WorkingHourInput:
      type: object
      properties:
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        isEnabled:
          type: boolean
      required: [dayOfWeek, startTime, endTime, isEnabled]
    BreakTime:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        label:
          type: ["string", "null"]
      required: [id, businessId, dayOfWeek, startTime, endTime, label]
    SetBreakTimesBody:
      type: object
      properties:
        breaks:
          type: array
          items:
            $ref: "#/components/schemas/BreakTimeInput"
      required: [breaks]
...
    AppointmentDetail:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        durationMinutes:
          type: integer
        clientName:
          type: string
        phoneNumber:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - businessId
        - serviceId
        - serviceName
        - durationMinutes
        - clientName
        - phoneNumber
        - appointmentDate
        - appointmentTime
        - notes
        - status
        - createdAt
...
    BusinessStats:
      type: object
      properties:
        totalAppointments:
          type: integer
        todayCount:
          type: integer
        thisWeekCount:
          type: integer
        upcomingCount:
          type: integer
      required: [totalAppointments, todayCount, thisWeekCount, upcomingCount]
    CustomerRecord:
      type: object
      properties:
        clientName:
          type: string
        phoneNumber:
          type: string
        totalVisits:
          type: integer
        totalRevenue:
          type: integer
          description: Total revenue in agorot
        lastVisitDate:
          type: string
        firstVisitDate:
          type: string
      required: [clientName, phoneNumber, totalVisits, totalRevenue, lastVisitDate, firstVisitDate]
...